scala 2.13.2

+++ server/src/main/scala/com/github/thebridsk/bridge/server/util/SystemTimeJVM.scala
@@ -6,7 +6,7 @@ object SystemTimeJVM {
 
   def apply() =
     SystemTime.setTimekeeper(new SystemTime {
-      def currentTimeMillis() = System.currentTimeMillis()
+      def currentTimeMillis() = System.currentTimeMillis().toDouble
     })
 
 }
